+++ cli-tool/templates/game_template.rb
@@ -10,12 +10,12 @@ def game_template
     Audio.open
 
     # Get the current monitor frame rate and set our target framerate to match.
-    set_target_fps(get_monitor_refresh_rate(get_current_monitor))
+    Window.target_fps = get_monitor_refresh_rate(get_current_monitor)
 
     # Define your main method
     def main
       # Get the amount of time passed since the last frame was rendered
-      delta = get_frame_time
+      delta = Window.frame_time
 
       # Your update logic goes here